# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Badge System API
  description: |
    REST API per il Badge System - SaaS multi-tenant per il tracciamento delle presenze nel retail.

    ## Autenticazione
    Tutti gli endpoint protetti richiedono JWT Bearer token nel header Authorization.

    ## Rate Limiting
    Limite: 100 richieste/minuto per IP. Headers di risposta includeranno X-RateLimit-*.

    ## Timezone
    Tutti i timestamp sono in ISO 8601 UTC (es. 2026-05-28T14:30:00Z)
  version: 1.0.0
  contact:
    name: Badge System Support
    url: https://badge.dataxiom.it/support
  license:
    name: Proprietary
    url: https://badge.dataxiom.it/license
servers:
  - url: https://api.badge.dataxiom.it
    description: Production
  - url: http://localhost:3000
    description: Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          example: emp_001
        email:
          type: string
          format: email
          example: mario@retail.it
        name:
          type: string
          example: Mario Rossi
        role:
          type: string
          enum: [employee, manager, admin]
          example: employee
      required: [id, email, name, role]
    CheckIn:
      type: object
      properties:
        id:
          type: string
          example: checkin_abc123
        employee_id:
          type: string
          example: emp_001
        employee_email:
          type: string
          format: email
        site_id:
          type: string
          example: site_milano_001
        timestamp:
          type: string
          format: date-time
          example: "2026-05-28T09:15:00Z"
        type:
          type: string
          enum: [IN, OUT]
          example: IN
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
          description: User ID who registered the check-in
          example: emp_001
        corrected_at:
          type: string
          format: date-time
          nullable: true
        corrected_by:
          type: string
          nullable: true
          description: User ID who corrected the check-in
          example: mgr_005
        status:
          type: string
          enum: [confirmed, corrected]
          example: confirmed
      required: [id, employee_id, site_id, timestamp, type, created_at, status]
...
    Meta:
      type: object
      properties:
        request_id:
          type: string
          example: req_xyz789
        timestamp:
          type: string
          format: date-time
        processing_ms:
          type: integer
          example: 85
      required: [request_id, timestamp, processing_ms]
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          example: INVALID_QR_CODE
        message:
          type: string
          example: QR code not found
        details:
          type: object
          example: { qr_code: "Invalid format" }
        http_status:
          type: integer
          example: 400
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorDetail'
        meta:
          $ref: '#/components/schemas/Meta'
      required: [success, error, meta]
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 47
        total_pages:
          type: integer
          example: 3
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Login con email e password
      description: Ottieni access token e refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required: [email, password]
            example:
              email: mario@retail.it
              password: secure_password_123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      refresh_token:
                        type: string
                      token_type:
                        type: string
                        example: Bearer
                      expires_in:
                        type: integer
                        example: 1800
                      user:
                        $ref: '#/components/schemas/User'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required: [refresh_token]
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      token_type:
                        type: string
                      expires_in:
                        type: integer
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Logout (invalidate token)
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Logged out successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════════════════════════════
  # CHECK-IN ENDPOINTS
  # ═══════════════════════════════════════════════════════════════════
...
  /api/v1/checkins:
    get:
      tags: [Check-ins]
      summary: List user's check-ins
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Items per page (max 100)
        - name: from_date
          in: query
          schema:
            type: string
            format: date
            example: "2026-05-01"
        - name: to_date
          in: query
          schema:
            type: string
            format: date
            example: "2026-05-31"
        - name: type
          in: query
          schema:
            type: string
            enum: [IN, OUT]
        - name: sort
          in: query
          schema:
            type: string
            enum: [timestamp_asc, timestamp_desc]
            default: timestamp_desc
      responses:
        '200':
          description: List of check-ins
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CheckIn'
                  pagination:
                    $ref: '#/components/schemas/PaginationMeta'
                  meta:
                    $ref: '#/components/schemas/Meta'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/export/csv:
    get:
      tags: [Reporting]
      summary: Export presences to CSV
      security:
        - BearerAuth: []
      parameters:
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
        - name: site_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════════════════════════════
  # ADMIN ENDPOINTS
  # ═══════════════════════════════════════════════════════════════════
...